it mIGHT WORK!!!!!!!!!! omg omg omg omg

+++ src/main/java/org/team639/lib/Constants.java
@@ -153,7 +153,7 @@ public interface Constants
         public static final double shooterD = 0.0002;
         public static final double shooterFF = 0.00018;
 
-        public static final TreeMap<Integer, AngleSpeed> shootMap = new TreeMap<>();
+        public static final TreeMap<Double, AngleSpeed> shootMap = new TreeMap<>();
 
 
     }
